smartinsight/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
whitenoise==6.9.0
dj-database-url==2.2.0
psycopg[binary]==3.2.6
pandas
scikit-learn
prophet
